requirements.txt
numpy
matplotlib
pandas
biopython
pyyaml